feat: add classes cart and goodslist
